Change parameter type from int to Long in predecirPorDni

+++ src/main/java/com/ProyectoDeAula5/Proyecto5/controller/PrediccionController.java
@@ -30,7 +30,7 @@ public class PrediccionController {
     }
 
     @GetMapping("/dni/{dni}")
-    public ResponseEntity<?> predecirPorDni(@PathVariable int dni) {
+    public ResponseEntity<?> predecirPorDni(@PathVariable Long dni) {
         try {
             Map<String, Object> resultado = servicio.predecirPorDni(dni);
             return ResponseEntity.ok(resultado);
